mybd/requirements.txt
# Core Flask dependencies
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
Werkzeug==3.1.3

# PostgreSQL database connector
psycopg2-binary==2.9.10

# SQLAlchemy
SQLAlchemy==2.0.43

# Utilities
tabulate==0.9.0

# Flask dependencies (will be installed automatically, but explicit for clarity)
click==8.1.8
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
blinker==1.9.0

# Python 3.12 compatibility
typing_extensions==4.15.0
cryptography==41.0.7

# Production server
gunicorn==21.2.0